Read on Toolbars, Menubars and Dialogs
Just more widgets, nothing serious. To make things clear, I'm not
targetting to be a master PyQt5-er or anything close, I just thought it
would be nice to be able to cook up GUIs and package them once in a
while to make my applications easier to use. Therefore, today's read was
not thorough and it will never be unless needed. For the 3 new widgets
I've seen today I've only studied them enough for confort and usability.

Happy Easter.