Basic Ray tracing redo
First version of the ray tracing program was messy and very buggy mainly
due to a lack of understanding, the new version is a bit cleaner with
the knowledge of a few assumptions that were made in version one that
were unknown to me. Adding camera movement is a bit heavy as that would
need that heavy linear algebra which am not in the mood to handle but
interested to consider some other time. Turns out ray tracing is very
expensive compute-wise thus low resolution in version two as the
default.